Moved smvspatial.c to the tools directory.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@918 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_image.c
@@ -291,7 +291,7 @@ mx_image_alloc( MX_IMAGE_FRAME **frame,
     ("%s: Invoked as mx_image_alloc( %p, %ld, %ld, %ld, %ld, %g, %ld, %ld )",
 		fname, frame, row_framesize, column_framesize,
 		image_format, byte_order, bytes_per_pixel,
-		header_length, image_length ));
+		(long) header_length, (long) image_length ));
 #endif
 
 	/* We either reuse an existing MX_IMAGE_FRAME or create a new one. */
@@ -479,10 +479,12 @@ mx_image_alloc( MX_IMAGE_FRAME **frame,
 		fname, (*frame), (*frame)->header_data, (*frame)->image_data ));
 
 	MX_DEBUG(-2,("%s: header_length = %ld, allocated_header_length = %ld",
-	    fname, (*frame)->header_length, (*frame)->allocated_header_length));
+		fname, (long) (*frame)->header_length,
+		(long) (*frame)->allocated_header_length));
 
 	MX_DEBUG(-2,("%s: image_length = %ld, allocated_image_length = %ld",
-	    fname, (*frame)->image_length, (*frame)->allocated_image_length));
+		fname, (long) (*frame)->image_length,
+		(long) (*frame)->allocated_image_length));
 #endif
 	return MX_SUCCESSFUL_RESULT;
 }
